refactor: AddCardContext 를 전역인 App 컴포넌트로 이동
