+++ src/App.jsx
@@ -1,26 +1,46 @@
 import './css/App.css';
-import React, { useState } from 'react';
+import React, { useState, useMemo } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import AddCard from './addCard/components/AddCard';
 import AddCardSuccess from './addSuccess/components/AddCardSuccess';
 import CardList from './cardList/components/CardList';
+import { Card as CardConstructor } from './util';
+import AddCardContext from './AddCardContext';
 
 function App() {
   const [cardList, setCardList] = useState([]);
+  const [card, setCard] = useState(CardConstructor());
 
-  const addCard = (card) => {
-    setCardList((prevList) => [...prevList, card]);
+  const updateCard = (name, value) => {
+    setCard((prevCard) => {
+      return { ...prevCard, [name]: value };
+    });
   };
+
+  const initCard = () => {
+    Object.keys(CardConstructor()).forEach((cardKey) => {
+      updateCard(cardKey, '');
+    });
+  };
+
+  const addCard = (newCard) => {
+    setCardList((prevList) => [...prevList, newCard]);
+    initCard();
+  };
+
+  const contextValue = useMemo(() => ({ card, updateCard }), [card]);
 
   return (
     <div className="App">
-      <Router basename={process.env.PUBLIC_URL}>
-        <Routes>
-          <Route path="/" element={<CardList cardList={cardList} />} />
-          <Route path="/addCard" element={<AddCard />} />
-          <Route path="/success" element={<AddCardSuccess addCard={addCard} />} />
-        </Routes>
-      </Router>
+      <AddCardContext.Provider value={contextValue}>
+        <Router basename={process.env.PUBLIC_URL}>
+          <Routes>
+            <Route path="/" element={<CardList cardList={cardList} />} />
+            <Route path="/addCard" element={<AddCard />} />
+            <Route path="/success" element={<AddCardSuccess addCard={addCard} />} />
+          </Routes>
+        </Router>
+      </AddCardContext.Provider>
     </div>
   );
 }
